requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pyaml
requests
django-cors-headers

